+++ lib/board/PrimaryBoard.cpp
@@ -19,7 +19,7 @@ PrimaryBoard::PrimaryBoard(std::shared_ptr<NcHandler> ncHandler)
                    ncHandler->get_default_channels(), PrimaryBoard::_symbols) {
     _cells = std::array<LeafBoard *, 9>();
 
-    for (int i = 0; i < 9; i++) {
+    for (unsigned int i = 0; i < 9; i++) {
         ncplane *plane = mGBoard.get_child_planes().at(i);
 
         LeafBoard *newBoard = new LeafBoard(ncHandler, plane);
@@ -33,7 +33,7 @@ PrimaryBoard::~PrimaryBoard() {
     }
 }
 
-CellOwner PrimaryBoard::get_cell_owner(const int INDEX) const {
+CellOwner PrimaryBoard::get_cell_owner(const unsigned int INDEX) const {
     return _cells.at(INDEX)->get_winner();
 }
 
@@ -47,7 +47,8 @@ void PrimaryBoard::draw_o(const unsigned int INDEX) {
     this->Board::draw_o(INDEX);
 }
 
-std::optional<LeafBoard *> PrimaryBoard::select_board(const int INDEX) const {
+std::optional<LeafBoard *>
+PrimaryBoard::select_board(const unsigned int INDEX) const {
     LeafBoard *cell = _cells.at(INDEX);
     std::optional<LeafBoard *> opt = std::optional(cell);
 
@@ -57,19 +58,19 @@ std::optional<LeafBoard *> PrimaryBoard::select_board(const int INDEX) const {
 ncplane_options def_primary_nopts(std::shared_ptr<NcHandler> ncHandler) {
     ncplane *std = ncHandler->get_stdplane();
 
-    const int ROWS = (3 * 11) + 2;
-    const int COLS = (3 * 20) + 2;
+    const unsigned int ROWS = (3u * 11u) + 2u;
+    const unsigned int COLS = (3u * 20u) + 2u;
 
     unsigned int std_rows, std_cols;
     ncplane_dim_yx(std, &std_rows, &std_cols);
 
     unsigned int std_center_y, std_center_x;
-    std_center_y = std_rows / 2;
-    std_center_x = std_cols / 2;
+    std_center_y = std_rows / 2u;
+    std_center_x = std_cols / 2u;
 
     int board_origin_y, board_origin_x;
-    board_origin_y = std_center_y - (ROWS / 2);
-    board_origin_x = std_center_x - (COLS / 2);
+    board_origin_y = static_cast<int>(std_center_y - (ROWS / 2u));
+    board_origin_x = static_cast<int>(std_center_x - (COLS / 2u));
 
     ncplane_options nopts = {
         board_origin_y, board_origin_x, ROWS, COLS, NULL, NULL, NULL, 0, 0, 0,
